refactor: rename Salesforce occurence with sfdx
