+++ test/fixtures/strategies/sfdx/sfdx-project.json
@@ -5,5 +5,5 @@
       "versionNumber": "1.0.0.NEXT"
     }
   ],
-  "name": "salesfore-test-repo"
+  "name": "sfdx-test-repo"
 }
